Basics of Person Model
Very basic version of the person. Meant to go at the very beginning of
the game to help set up back story. Hierarchical model of the person,
with animating arms.

Also, Clicking in the dialogue box is going to progress the dialogue.
Then on the final click, it could go into the animation outside the
house of him getting in the car.